requirements.txt
moviepy
pika
python-dotenv
psycopg2
requests
elevenlabs
boto3
langchain
langchain_core
langchain_openai